requirements.txt
asgiref==3.8.1
Django==5.2.3
django-extensions==4.1
djangorestframework==3.16.0
environs==14.2.0
iniconfig==2.1.0
marshmallow==4.0.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
sqlparse==0.5.3
environs==14.2.0